update
git-svn-id: http://svn.tudos.org/repos/tudos/trunk@308 a704ac0b-3a55-4d43-a2a9-7be6f07c34fb

+++ l4/pkg/uclibc/lib/contrib/uclibc/libc/sysdeps/linux/common/ioctl.c
@@ -7,7 +7,7 @@
  * Licensed under the LGPL v2.1, see the file COPYING.LIB in this tarball.
  */
 
-#include "syscalls.h"
+#include <sys/syscall.h>
 #include <stdarg.h>
 #include <sys/ioctl.h>
 
@@ -15,7 +15,7 @@ libc_hidden_proto(ioctl)
 
 #define __NR___syscall_ioctl __NR_ioctl
 static inline
-_syscall3(int, __syscall_ioctl, int, fd, unsigned long int, request, void *, arg);
+_syscall3(int, __syscall_ioctl, int, fd, int, request, void *, arg);
 
 int ioctl(int fd, unsigned long int request, ...)
 {
@@ -24,8 +24,8 @@ int ioctl(int fd, unsigned long int request, ...)
 
     va_start(list, request);
     arg = va_arg(list, void *);
-
     va_end(list);
+
     return __syscall_ioctl(fd, request, arg);
 }
 libc_hidden_def(ioctl)
